Feat: 자신이 쓴 글만 삭제, 수정 가능하게 변경

+++ ott_posts/views.py
@@ -1,10 +1,10 @@
 from django.shortcuts import render, redirect
 from .models import Ott
 from .forms import OttForm
-from django.http import HttpResponseRedirect
-from django.urls import reverse
+from django.http import HttpResponseRedirect, HttpResponseForbidden
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 from django.shortcuts import get_object_or_404
+from django.contrib.auth.decorators import login_required
 
 def ott_list(request):
     otts = Ott.objects.order_by('-pk')
@@ -25,28 +25,41 @@ def ott_detail(request, ott_id):
     ott = Ott.objects.get(id=ott_id)
     return render(request, 'ott_posts/ott_detail.html', {'ott': ott})
 
+@login_required
 def ott_create(request):
     if request.method == 'POST':
         form = OttForm(request.POST)
         if form.is_valid():
-            new_item = form.save()
+            new_item = form.save(commit=False)
+            new_item.author = request.user  # 사용자 정보를 글과 연결
+            new_item.save()
             return HttpResponseRedirect('/ott/')
     form = OttForm()
     return render(request, 'ott_posts/ott_form.html', {'form': form})
 
 
+@login_required
 def ott_delete(request):
     ott_id = request.GET.get('id')
-    
+
     if ott_id:
         item = get_object_or_404(Ott, pk=ott_id)
+
+        # 사용자가 글의 작성자인지 확인
+        if request.user != item.author:
+            return HttpResponseForbidden("글을 삭제할 권한이 없습니다.")
+        
         item.delete()
         
     return redirect('ott_list')
 
+@login_required
 def ott_update(request, id):
     item = get_object_or_404(Ott, pk=id)
 
+    if request.user != item.author:
+        return HttpResponseForbidden("글을 수정할 권한이 없습니다.")
+    
     if request.method == 'POST':
         form = OttForm(request.POST, instance=item)
         if form.is_valid():
@@ -57,3 +70,4 @@ def ott_update(request, id):
         form = OttForm(instance=item)
 
     return render(request, 'ott_posts/ott_update.html', {'form': form})
+
